requirements.txt
aiogram==3.15.0
yt-dlp==2024.12.13
python-dotenv==1.0.1
aiofiles==24.1.0
